refactor componenets into folders
